backend:  - added auth & permission module MD doc files  - code clean-up & refactoring - updated backend MD: project structure & app launch  - applied Authorize decorator usage examples is UserController
