+++ packages/backend/src/modules/auth/dto/index.ts
@@ -2,5 +2,4 @@ export * from "./user-registration.dto";
 export * from "./acl-dashboard.dto";
 export * from "./acl.dto";
 export * from "./role-permission.dto";
-export * from "./permission.dto";
 export * from "./role.dto";
