+++ packages/backend/src/modules/auth/permission/dto/index.ts
@@ -1,0 +1,1 @@
+export * from "./permission.dto";
