+++ packages/backend/src/modules/auth/permission/dto/permission.dto.ts
@@ -1,5 +1,5 @@
-import { BaseDto } from "../../../base.dto";
-import { PermissionName } from "../../db";
+import { BaseDto } from "../../../../base.dto";
+import { PermissionName } from "../../../db";
 
 
 export class PermissionDto extends BaseDto {
