+++ packages/backend/src/modules/auth/permission/index.ts
@@ -1,3 +1,4 @@
+export * from "./dto";
 export * from "./services";
 export * from "./permission.di";
 export * from "./permission.module";
