+++ packages/backend/src/modules/auth/permission/permission.module.md
@@ -1,0 +1,10 @@
+# Permission Module
+
+Provides Permission utilities for application.
+
+Features:
+ - **Permission Service** - Core logic. Exposes permission management functionality
+
+ For setting up new permissions, you need to:
+ - Update PermissionName enum
+ - add new permission via PermissionMigrator. Examples can be found in Migrator documentation
